requirements.txt
# Core requirements for SDR AI Agent

# LangChain and AI (Core dependencies)
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.20
langgraph>=0.2.0
langsmith>=0.1.17,<0.2.0
langchain-ollama>=0.2.0

# MCP Client for BrightData integration
mcp==1.12.0

# HTTP client for MCP and API calls
httpx>=0.27.0
aiohttp>=3.9.1

# Environment configuration
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: Code formatting (development only)
# black==23.11.0
# isort==5.12.0
# mypy==1.7.1